requirements.txt
python-dotenv

fastapi
uvicorn
python-multipart

ollama
openai
langgraph
google-genai
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
pytz>=2024.1

discord.py

huggingface_hub
onnxruntime-gpu==1.20.1
kokoro-onnx

# CUDA libraries for GPU acceleration (TTS)
nvidia-cublas-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cuda-nvrtc-cu12
sentencepiece
librosa
soundfile
resampy
transformers
torch

discord-ext-voice-recv>=0.5.0a167
PyNaCl>=1.5.0
scipy>=1.10.0
httpx
pymupdf4llm

# Performance Optimization (Speech Pipeline)
onnx-asr[gpu,hub]>=0.10.0
soxr>=0.3.0

# Long-term Memory (Semantic Search)
sentence-transformers>=2.2.0
chromadb>=0.4.0